Organize breakpoints and theme under style
