+++ src/style/breakpoints.js
@@ -1,4 +1,4 @@
-export const BREAKPOINTS = {
+export const breakpoints = {
   min: 360,
   small: 680,
   medium: 720,
