+++ src/style/theme.js
@@ -1,4 +1,4 @@
-export default {
+export const theme = {
   _name: 'network-dashboard',
   _appearance: 'light',
 }
